Split export import flow
